requirements.txt
playwright
streamlit
python-dotenv
pytest
